# docs/swagger.yaml
...
basePath: /api/v1
definitions:
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.IATA:
    properties:
      displayName:
        type: string
      iata:
        type: string
      lat:
        type: number
      lon:
        type: number
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.Packet:
    properties:
      channelHash:
        description: hex-encoded single byte; non-nil for group_text/group_data
        type: string
      decrypted:
        description: true if group text was successfully decrypted
        type: boolean
      firstHeardAt:
        description: epoch ms
        type: integer
      firstToLastMs:
        description: ms between first and last observation
        type: integer
      header:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader'
      lastHeardAt:
        description: epoch ms
        type: integer
      observationCount:
        type: integer
      observations:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationDetail'
        type: array
      originPubkey:
        description: hex-encoded; nil when not extractable from payload
        type: string
      packetHash:
        type: string
      parsedPayload:
        items:
          type: integer
        type: array
      rawPayload:
        description: hex-encoded payload bytes (excludes header and path)
        type: string
      resolvedRoute:
        description: 'trace packets only: resolved intended route'
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      scope:
        description: matched transport scope name e.g. "#bc"
        type: string
      transportCodes:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes'
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader:
    properties:
      payloadType:
        description: bits 2-5
        type: integer
      payloadTypeName:
        description: advert, request, group_text, etc.
        type: string
      payloadVersion:
        description: bits 6-7
        type: integer
      raw:
        description: hex-encoded single byte
        type: string
      routeType:
        description: bits 0-1
        type: integer
      routeTypeName:
        description: FLOOD, DIRECT, TRANSPORT_FLOOD, TRANSPORT_DIRECT
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketLatestObserver:
    properties:
      displayName:
        type: string
      iata:
        type: string
      id:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationDetail:
    properties:
      heardAt:
        description: epoch ms
        type: integer
      iata:
        type: string
      id:
        type: integer
      observerId:
        type: string
      observerName:
        type: string
      pathBytes:
        description: hex-encoded accumulated path hashes
        type: string
      pathLength:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength'
      propagationTimeMs:
        description: ms since first observation; 0 for first
        type: integer
      radio:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketRadio'
      resolvedPath:
        description: per-observation resolved path hashes
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      rssi:
        type: integer
      snr:
        type: number
      sourceBroker:
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength:
    properties:
      hashSize:
        description: per-hop hash size in bytes (1, 2, or 3)
        type: integer
      hopCount:
        description: number of path hashes present
        type: integer
      raw:
        description: hex-encoded single byte
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketRadio:
    properties:
      bandwidthKhz:
        type: number
      codingRate:
        type: integer
      freqMhz:
        type: number
      spreadFactor:
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketSummary:
    properties:
      firstHeardAt:
        description: epoch ms
        type: integer
      lastHeardAt:
        description: epoch ms
        type: integer
      latestObserver:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketLatestObserver'
      observationCount:
        type: integer
      packetHash:
        description: hex-encoded
        type: string
      payloadType:
        type: integer
      payloadTypeName:
        type: string
      routeType:
        type: integer
      routeTypeName:
        type: string
      scope:
        description: matched transport scope name e.g. "#bc"
        type: string
      summary:
        description: human-readable payload summary
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes:
    properties:
      regionCode:
        type: integer
      subRegionCode:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.Region:
    properties:
      centerLat:
        description: map center latitude
        type: number
      centerLng:
        description: map center longitude
        type: number
      description:
        type: string
      iatas:
        description: member IATA codes
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
      slug:
        description: URL-safe identifier e.g. "western-canada"
        type: string
      zoomLevel:
        description: suggested map zoom level
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.RegionSummary:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        description: URL-safe identifier e.g. "western-canada"
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop:
    properties:
      confidence:
        description: '"high", "ambiguous", or "none"'
        type: string
      nodes:
        description: empty for "none", one for "high", multiple for "ambiguous"
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        type: array
      snr:
        type: number
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode:
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      publicKey:
        description: hex-encoded prefix used for resolution
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeDetail:
    properties:
      iataCount:
        type: integer
      iatas:
        items:
          type: string
        type: array
      name:
        type: string
      nodeCount:
        type: integer
      observerCount:
        type: integer
      packetCount:
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeStats:
    properties:
      name:
        description: normalized scope name e.g. "#bc"
        type: string
      nodeCount:
        description: distinct nodes with this as their default scope
        type: integer
      observerCount:
        description: distinct observers that forwarded packets in this scope
        type: integer
      packetCount:
        description: distinct packets matched to this scope
        type: integer
    type: object
...
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: MIT
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: "1.0"
paths:
...
  /iatas:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.IATA'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List all IATA codes
      tags:
      - IATAs
  /iatas/{iata}:
    get:
      parameters:
      - description: 3-letter IATA code
        in: path
        name: iata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.IATA'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get a single IATA code
      tags:
      - IATAs
...
  /packets:
    get:
      parameters:
      - description: Filter by payload type integer
        in: query
        name: payloadType
        type: integer
      - description: Filter by payload type name (advert, grp_txt, txt_msg, trace,
          anon_req)
        in: query
        name: payloadTypeName
        type: string
      - description: Filter by route type (0=transport_flood, 1=flood, 2=direct, 3=transport_direct)
        in: query
        name: routeType
        type: integer
      - description: Filter by single IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: Filter by first_heard_at >= since (epoch ms)
        in: query
        name: since
        type: integer
      - description: Filter by first_heard_at <= until (epoch ms)
        in: query
        name: until
        type: integer
      - description: last_heard_at epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List packets
      tags:
      - Packets
  /packets/{packetHash}:
    get:
      parameters:
      - description: Packet hash (hex)
        in: path
        name: packetHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Packet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get full packet detail. For trace packets (payloadType=9), includes
        resolvedRoute
      tags:
      - Packets
  /packets/backfill:
    get:
      parameters:
      - description: Return packets with observations after this ID (use last WS event
          observation ID)
        in: query
        name: afterObservationId
        required: true
        type: integer
      - description: Filter by payload type integer
        in: query
        name: payloadType
        type: integer
      - description: Filter by payload type name
        in: query
        name: payloadTypeName
        type: string
      - description: Filter by route type
        in: query
        name: routeType
        type: integer
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      - description: Filter by transport scope name
        in: query
        name: scope
        type: string
      - description: Max results (default 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketSummary'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Backfill packets after a given observation ID
      tags:
      - Packets
  /regions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RegionSummary'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List all regions
      tags:
      - Regions
  /regions/{regionId}:
    get:
      parameters:
      - description: Region ID
        in: path
        name: regionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Region'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get a single region
      tags:
      - Regions
...
  /scopes:
    get:
      parameters:
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List transport scopes
      tags:
      - Scopes
  /scopes/{name}:
    get:
      parameters:
      - description: 'Scope name e.g. %23bc (URL-encoded #bc)'
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get scope detail by name
      tags:
      - Scopes
...
  /stats/scopes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Scope statistics
      tags:
      - Stats
...
schemes:
- http
- https
swagger: "2.0"
...